move CODEOWNER logic into dedicated tasks/libs/owners/linter.py file
