+++ tasks/github_tasks.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import os
 import re
-import sys
 import time
 from collections import Counter
 from functools import lru_cache
@@ -19,11 +18,11 @@ from tasks.libs.ciproviders.github_actions_tools import (
     print_workflow_conclusion,
     trigger_macos_workflow,
 )
-from tasks.libs.common.color import color_message
 from tasks.libs.common.constants import DEFAULT_BRANCH, DEFAULT_INTEGRATIONS_CORE_BRANCH
 from tasks.libs.common.datadog_api import create_gauge, send_metrics
 from tasks.libs.common.junit_upload_core import repack_macos_junit_tar
 from tasks.libs.common.utils import get_git_pretty_ref
+from tasks.libs.owners.linter import codeowner_has_orphans, directory_has_packages_without_owner
 from tasks.libs.owners.parsing import read_owners
 from tasks.libs.pipeline.notifications import GITHUB_SLACK_MAP
 from tasks.release import _get_release_json_value
@@ -143,7 +142,7 @@ def lint_codeowner(_, owners_file=".github/CODEOWNERS"):
     owners = read_owners(owners_file)
 
     # Define linters
-    linters = [_has_packages_without_owner, _has_orphans_in_codeowner]
+    linters = [directory_has_packages_without_owner, codeowner_has_orphans]
 
     # Execute linters
     for linter in linters:
@@ -151,106 +150,6 @@ def lint_codeowner(_, owners_file=".github/CODEOWNERS"):
             exit_code = 1
 
     raise Exit(code=exit_code)
-
-
-def _has_packages_without_owner(owners, folder="pkg"):
-    """
-    Check every package in `pkg` has an owner
-    """
-
-    error = False
-
-    for x in os.listdir(folder):
-        path = os.path.join("/" + folder, x)
-        if all(owner[1].rstrip('/') != path for owner in owners.paths):
-            if not error:
-                print(
-                    color_message("The following packages don't have owner in CODEOWNER file", "red"), file=sys.stderr
-                )
-                error = True
-            print(color_message(f"\t- {path}", "orange"), file=sys.stderr)
-
-    return error
-
-
-def _has_orphans_in_codeowner(owners):
-    """
-    Check that every rule in codeowners file point to an existing file/directory
-    """
-
-    err_invalid_rule_path = False
-    err_orphans_path = False
-
-    for rule in owners.paths:
-        try:
-            # Get the static part of the rule path, removing matching subpath (such as '*')
-            static_root = _get_static_root(rule[1])
-        except Exception:
-            err_invalid_rule_path = True
-            print(
-                color_message(
-                    f"[UNSUPPORTED] The following rule's path does not start with '/' anchor: {rule[1]}", "red"
-                ),
-                file=sys.stderr,
-            )
-            continue
-
-        if not _is_pattern_in_fs(static_root, rule[0]):
-            if not err_orphans_path:
-                print(
-                    color_message(
-                        "The following rules are outdated: they don't point to existing file/directory", "red"
-                    ),
-                    file=sys.stderr,
-                )
-                err_orphans_path = True
-            print(color_message(f"\t- {rule[1]}\t{rule[2]}", "orange"), file=sys.stderr)
-
-    return err_invalid_rule_path or err_orphans_path
-
-
-def _get_static_root(pattern):
-    """
-    _get_static_root returns the longest prefix path from the pattern without any wildcards.
-    """
-    result = "."
-
-    if not pattern.startswith("/"):
-        raise Exception()
-
-    # We remove the '/' anchor character from the path
-    pattern = pattern[1:]
-
-    for elem in pattern.split("/"):
-        if '*' in elem:
-            return result
-        result = os.path.join(result, elem)
-    return result
-
-
-def _is_pattern_in_fs(path, pattern):
-    """
-    Checks if a given pattern matches any file within the specified path.
-
-    Args:
-        path (str): The file or directory path to search within.
-        pattern (re.Pattern): The compiled regular expression pattern to match against file paths.
-
-    Returns:
-        bool: True if the pattern matches any file path within the specified path, False otherwise.
-    """
-    if os.path.isfile(path):
-        return True
-    elif os.path.isdir(path):
-        for root, _, files in os.walk(path):
-            for name in files:
-                # file_path is the relative path from the root of the repo, without "./" at the begining
-                file_path = os.path.join(root, name)[2:]
-
-                # Check if the file path matches any of the regex patterns
-                if pattern.match(file_path):
-                    return True
-    return False
 
 
 @task
